# Optimize zero reads