@nirs 2024-10-19
@AkihiroSuda current version is simpler, needs no temporary buffer, and much faster.